+++ libraries/IRremote/src/IRProtocol.h
@@ -38,10 +38,13 @@
  */
 typedef enum {
     UNKNOWN = 0,
+    PULSE_DISTANCE,
+    PULSE_WIDTH,
     DENON,
     DISH,
     JVC,
     LG,
+    LG2,
     NEC,
     PANASONIC,
     KASEIKYO,
@@ -54,6 +57,7 @@ typedef enum {
     SAMSUNG,
     SHARP,
     SONY,
+    ONKYO,
     APPLE,
     BOSEWAVE,
     LEGO_PF,
@@ -67,12 +71,25 @@ const __FlashStringHelper* getProtocolString(decode_type_t aProtocol);
 #define PROTOCOL_IS_MSB_FIRST true
 
 /*
+ * Carrier frequencies for various protocols
+ */
+#define SONY_KHZ        40
+#define BOSEWAVE_KHZ    38
+#define DENON_KHZ       38
+#define JVC_KHZ         38
+#define LG_KHZ          38
+#define NEC_KHZ         38
+#define SAMSUNG_KHZ     38
+#define KASEIKYO_KHZ    37
+#define RC5_RC6_KHZ     36
+
+/*
  * Constants for some protocols
  */
 #define PANASONIC_VENDOR_ID_CODE    0x2002
-#define SHARP_VENDOR_ID_CODE        0x5AAA
 #define DENON_VENDOR_ID_CODE        0x3254
 #define MITSUBISHI_VENDOR_ID_CODE   0xCB23
+#define SHARP_VENDOR_ID_CODE        0x5AAA
 #define JVC_VENDOR_ID_CODE          0x0103
 
 #define SIRCS_12_PROTOCOL       12